requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mystbin-py==7.0.2
pillow==11.1.0